ccsr: split into modules
The main file was getting rather large. Split it into modules to make it
more manageable. This commit is pure code movement.
